+++ src/mpservice/__init__.py
@@ -21,4 +21,10 @@ The package `mpservice <https://github.com/zpz/mpservice>`_ provides utilities f
 """
 
 
-__version__ = "0.13.3b3"
+__version__ = "0.13.3"
+
+
+import builtins
+
+class TimeoutError(builtins.TimeoutError):
+    pass