+++ tests/test_mpserver.py
@@ -14,10 +14,10 @@ from mpservice.mpserver import (
     Server,
     SwitchServlet,
     ThreadServlet,
-    TimeoutError,
     Worker,
     make_worker,
 )
+from mpservice import TimeoutError
 from mpservice.multiprocessing import RemoteException, is_remote_exception
 from mpservice.streamer import Stream