+++ tests/test_multiprocessing.py
@@ -18,6 +18,8 @@ from mpservice.multiprocessing import (
     wait,
 )
 from mpservice.threading import Thread
+from mpservice import TimeoutError
+
 
 logger = logging.getLogger(__name__)
 
@@ -33,7 +35,7 @@ def delay_double(x, delay=2):
     raise SystemExit()
 
 
-def _test_thread_process(cls, TimeoutError):
+def _test_thread_process(cls):
     # No exception
     t = cls(target=delay_double, args=(3,))
     t.start()
@@ -88,11 +90,11 @@ def _test_thread_process(cls, TimeoutError):
 
 @pytest.mark.filterwarnings("ignore::pytest.PytestUnhandledThreadExceptionWarning")
 def test_thread():
-    _test_thread_process(Thread, mpservice.threading.TimeoutError)
+    _test_thread_process(Thread)
 
 
 def test_process():
-    _test_thread_process(Process, mpservice.multiprocessing.TimeoutError)
+    _test_thread_process(Process)
 
 
 def goo(x, q):